pull gRPC client implementations to own package
