+++ dfs/internal/coordinator/server.go
@@ -19,7 +19,7 @@ import (
 // 3. Reply with where the client should upload each chunk (primary + replicas).
 func (c *Coordinator) UploadFile(ctx context.Context, pb *proto.UploadRequest) (*proto.UploadResponse, error) {
 	// transform into internal representation
-	req := newUploadRequestFromProto(pb)
+	req := common.UploadRequestFromProto(pb)
 	c.logger.Info("Received UploadRequest")
 
 	// Calculate number of chunks needed
@@ -38,7 +38,7 @@ func (c *Coordinator) UploadFile(ctx context.Context, pb *proto.UploadRequest) (
 	}
 	c.logger.Debug("Selected nodes for chunk distribution", slog.Int("num_nodes", len(nodes)))
 
-	assignments := make([]ChunkLocation, numChunks)
+	assignments := make([]common.ChunkLocation, numChunks)
 	for i := 0; i < numChunks; i++ {
 		chunkID := common.FormatChunkID(req.Path, i)
 
@@ -54,7 +54,7 @@ func (c *Coordinator) UploadFile(ctx context.Context, pb *proto.UploadRequest) (
 		}
 
 		// Add chunk location to assignment
-		assignments[i] = ChunkLocation{
+		assignments[i] = common.ChunkLocation{
 			ChunkID: chunkID,
 			Nodes:   selectedNodes,
 		}
@@ -65,7 +65,7 @@ func (c *Coordinator) UploadFile(ctx context.Context, pb *proto.UploadRequest) (
 	go c.metadataManager.trackUpload(sessionID, req, numChunks)
 	c.logger.Debug("Created upload session", slog.String("session_id", sessionID), slog.String("file_path", req.Path))
 
-	return UploadResponse{
+	return common.UploadResponse{
 		ChunkLocations: assignments,
 		SessionID:      sessionID,
 	}.ToProto(), nil
@@ -81,7 +81,7 @@ func (c *Coordinator) DownloadFile(ctx context.Context, req *proto.DownloadReque
 	}
 
 	// Build chunk sources
-	chunkLocations := make([]ChunkLocation, 0, len(fileInfo.Chunks))
+	chunkLocations := make([]common.ChunkLocation, 0, len(fileInfo.Chunks))
 
 	// Find available nodes to download from for this chunk
 	for i, chunk := range fileInfo.Chunks {
@@ -91,14 +91,14 @@ func (c *Coordinator) DownloadFile(ctx context.Context, req *proto.DownloadReque
 			return nil, status.Error(codes.NotFound, "no available nodes")
 		}
 
-		chunkLocations[i] = ChunkLocation{
+		chunkLocations[i] = common.ChunkLocation{
 			ChunkID: chunk.Header.ID,
 			Nodes:   nodes,
 		}
 	}
 
 	c.logger.Debug("Replying to client with chunk locations", slog.Int("num_chunks", len(chunkLocations)))
-	return DownloadResponse{
+	return common.DownloadResponse{
 		FileInfo:       *fileInfo,
 		ChunkLocations: chunkLocations,
 		SessionID:      uuid.NewString(), // TODO: this should be a session id that is used to track the download
@@ -116,18 +116,18 @@ func (c *Coordinator) DownloadFile(ctx context.Context, req *proto.DownloadReque
 // }
 
 func (c *Coordinator) ConfirmUpload(ctx context.Context, pb *proto.ConfirmUploadRequest) (*proto.ConfirmUploadResponse, error) {
-	req := ConfirmUploadRequestFromProto(pb)
+	req := common.ConfirmUploadRequestFromProto(pb)
 	c.logger.Debug("Received ConfirmUploadRequest from client", slog.String("session_id", req.SessionID))
 
 	if err := c.metadataManager.commit(req.SessionID, req.ChunkInfos, c.metaStore); err != nil {
 		c.logger.Error("Failed to commit metadata", slog.String("error", err.Error()))
-		return ConfirmUploadResponse{
+		return common.ConfirmUploadResponse{
 			Success: false,
 			Message: "Failed to commit metadata",
 		}.ToProto(), nil
 	}
 
-	return ConfirmUploadResponse{
+	return common.ConfirmUploadResponse{
 		Success: true,
 	}.ToProto(), nil
 }
@@ -143,7 +143,7 @@ func (c *Coordinator) RegisterDataNode(ctx context.Context, pb *proto.RegisterDa
 	nodes, version := c.nodeManager.ListNodes()
 
 	c.logger.Debug("Registered datanode, replying with current node list and version", slog.Int("num_nodes", len(nodes)), slog.Int("version", int(version)))
-	return RegisterDataNodeResponse{
+	return common.RegisterDataNodeResponse{
 		Success:        true,
 		Message:        "Node registered successfully",
 		FullNodeList:   nodes,
@@ -153,13 +153,13 @@ func (c *Coordinator) RegisterDataNode(ctx context.Context, pb *proto.RegisterDa
 
 // DataNodes periodically communicate their status to the coordinator
 func (c *Coordinator) DataNodeHeartbeat(ctx context.Context, pb *proto.HeartbeatRequest) (*proto.HeartbeatResponse, error) {
-	req := HeartbeatRequestFromProto(pb)
+	req := common.HeartbeatRequestFromProto(pb)
 	c.logger.Debug("Received HeartbeatRequest from data node", slog.String("node_id", req.NodeID))
 
 	node, exists := c.nodeManager.GetNode(req.NodeID)
 	if !exists {
 		c.logger.Error("Data node not found", slog.String("node_id", req.NodeID))
-		return HeartbeatResponse{
+		return common.HeartbeatResponse{
 			Success: false,
 			Message: "node is not registered",
 		}.ToProto(), nil
@@ -171,7 +171,7 @@ func (c *Coordinator) DataNodeHeartbeat(ctx context.Context, pb *proto.Heartbeat
 	updates, currentVersion, err := c.nodeManager.GetUpdatesSince(req.LastSeenVersion)
 	if err != nil {
 		c.logger.Error("Failed to get updates since provided", slog.String("error", err.Error()))
-		return HeartbeatResponse{
+		return common.HeartbeatResponse{
 			Success:            true, // heartbeat was a success, as lastSeen & status were updated, but node requires resync
 			Message:            "version too old",
 			FromVersion:        req.LastSeenVersion,
@@ -181,7 +181,7 @@ func (c *Coordinator) DataNodeHeartbeat(ctx context.Context, pb *proto.Heartbeat
 	}
 
 	c.logger.Debug("Replying to data node with updates", slog.Int("num_updates", len(updates)))
-	return HeartbeatResponse{
+	return common.HeartbeatResponse{
 		Success:            true,
 		Message:            "ok",
 		Updates:            updates,
@@ -195,7 +195,7 @@ func (c *Coordinator) DataNodeHeartbeat(ctx context.Context, pb *proto.Heartbeat
 // ListNodes returns a list of all registered data nodes
 func (c *Coordinator) ListNodes(ctx context.Context, req *proto.ListNodesRequest) (*proto.ListNodesResponse, error) {
 	nodes, version := c.nodeManager.ListNodes()
-	return ListNodesResponse{
+	return common.ListNodesResponse{
 		Nodes:          nodes,
 		CurrentVersion: version,
 	}.ToProto(), nil
