+++ dfs/internal/datanode/types.go
@@ -4,13 +4,12 @@ import (
 	"context"
 	"log/slog"
 
+	"github.com/mochivi/distributed-file-system/internal/clients"
 	"github.com/mochivi/distributed-file-system/internal/cluster"
 	"github.com/mochivi/distributed-file-system/internal/common"
 	"github.com/mochivi/distributed-file-system/internal/storage"
 	"github.com/mochivi/distributed-file-system/pkg/logging"
 	"github.com/mochivi/distributed-file-system/pkg/proto"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/credentials/insecure"
 )
 
 // Implements the proto.DataNodeServiceServer interface
@@ -27,16 +26,9 @@ type DataNodeServer struct {
 	logger *slog.Logger
 }
 
-// Wrapper over the proto.DataNodeServiceClient interface
-type DataNodeClient struct {
-	client proto.DataNodeServiceClient
-	conn   *grpc.ClientConn
-	Node   *common.DataNodeInfo
-}
-
 type IReplicationManager interface {
-	paralellReplicate(clients []*DataNodeClient, chunkInfo common.ChunkHeader, data []byte, requiredReplicas int) ([]*common.DataNodeInfo, error)
-	replicate(ctx context.Context, client *DataNodeClient, req common.ChunkHeader, data []byte, clientLogger *slog.Logger) error
+	paralellReplicate(clients []*clients.DataNodeClient, chunkInfo common.ChunkHeader, data []byte, requiredReplicas int) ([]*common.DataNodeInfo, error)
+	replicate(ctx context.Context, client *clients.DataNodeClient, req common.ChunkHeader, data []byte, clientLogger *slog.Logger) error
 }
 
 type ISessionManager interface {
@@ -58,24 +50,4 @@ func NewDataNodeServer(store storage.ChunkStorage, replicationManager IReplicati
 		Config:             config,
 		logger:             datanodeLogger,
 	}
-}
-
-func NewDataNodeClient(node *common.DataNodeInfo, opts ...grpc.DialOption) (*DataNodeClient, error) {
-	opts = append(opts, grpc.WithTransportCredentials(insecure.NewCredentials()))
-
-	conn, err := grpc.NewClient(
-		node.Endpoint(),
-		opts...,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	client := proto.NewDataNodeServiceClient(conn)
-
-	return &DataNodeClient{
-		client: client,
-		conn:   conn,
-		Node:   node,
-	}, nil
 }
